add erlangen, distributed message passing for CCL.
author's words:

> Author here. It still has a long way to go. On the platform branch you can find the beginning of an OTP-like framework (notably includes a basic supervisor implementation). Also for Erlangen to become really useful we need more asynchronous libraries such as IOLib for common network protocols such as HTTP and database client protocols.

Caveat: the distribution features are currently implemented with Clozure's blocking socket API. This is a double edged sword, the main issue is that SEND to a remote agent can block for some time depending on message size and/or network timeouts (but not indefinitely). The secondary issue is that there is currently one thread per incoming node connection. Should probably use IOLib to rewrite the distribution layer in a more asynchronous design.

Also: there will be bugs, Erlangen is still missing a thorough integration test that exercises the distribution layer.

https://www.reddit.com/r/Common_Lisp/comments/5gjbqd/erlangen_brings_distributed_asynchronous_message/?utm_term=7327c7f8-340c-4df5-ba64-25d70ebc2fb0&utm_medium=search&utm_source=reddit&utm_name=Common_Lisp&utm_content=2