Keep CNAME and _config.yml when deploying to gh-pages
When pushing the dist/ director to gh-pages, Travis deletes any
existing files that aren't in dist. That includes the CNAME file that
points awesome-cl.com to the github pages sites.

This update clones the existing gh-branches into dist, then copies in
the new README, so that we don't lose any files existing on gh-pages.