Keep notabug repo in sync with GitHub
Fixes #193 (hopefully :D)

Notabug doesn't have a 'mirror' option, like GitHub does. But we're already running these commands in the Makefile when we merge to master, so we can push all new changes to notabug here also.

(FWIW - the Makefile for gh-pages is actually unnecessary, we should be able to read directly from `master`, but...if it ain't broke don't fix it! :wrench:)